requirements.txt
action-msgs==1.2.2
actionlib-msgs==4.9.0
altair==5.5.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
annotated-types==0.7.0
apturl==0.5.2
arabic-reshaper==3.0.0
argcomplete==1.8.1
asn1crypto==1.5.1
attrs==25.4.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
bloom==0.13.0
breezy==3.2.1
Brlapi==0.8.3
builtin-interfaces==1.2.2
cachetools==5.0.0
catkin-pkg==1.1.0
catkin-pkg-modules==1.1.0
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.20.1
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.2
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cryptography==43.0.3
cssselect2==0.8.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
diagnostic-msgs==4.9.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
duplicity==0.8.21
empy==3.3.4
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
flake8==4.0.1
freetype-py==2.5.1
future==0.18.2
geometry-msgs==4.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==1.7.11
google-auth==2.43.0
google-auth-httplib2==0.1.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gpg==1.16.0
greenlet==1.1.2
grpcio==1.76.0
grpcio-status==1.71.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
language-selector==0.1
lark==1.1.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.2
lockfile==0.12.2
louis==3.20.0
lxml==6.0.2
macaroonbakery==1.3.1
Markdown==3.10
MarkupSafe==3.0.3
mccabe==0.6.1
mercurial==6.1.1
message-filters==4.3.11
monotonic==1.6
more-itertools==8.10.0
narwhals==2.12.0
nav-msgs==4.9.0
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
osrf-pycommon==2.1.6
packaging==21.3
pandas==2.3.3
paramiko==2.9.3
patiencediff==0.2.1
pexpect==4.8.0
Pillow==9.0.1
plotly==6.5.0
pluggy==0.13.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.23
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydocstyle==6.1.1
PyDrive==1.3.1
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==2.4.7
pypdf==6.4.0
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.3
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.7
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==0.9.1
rlPyCairo==0.4.0
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosdep==0.26.0
rosdep-modules==0.26.0
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rospkg-modules==1.6.0
rpds-py==0.29.0
rpyutils==0.2.2
rsa==4.8
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six==1.16.0
smmap==5.0.2
snowballstemmer==2.2.0
soupsieve==2.3.1
SQLAlchemy==2.0.44
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
streamlit==1.51.0
svglib==1.6.0
systemd-python==234
tenacity==9.1.2
tf2-geometry-msgs==0.25.17
tf2-kdl==0.25.17
tf2-msgs==0.25.17
tf2-py==0.25.17
tf2-ros-py==0.25.17
tf2-sensor-msgs==0.25.17
tf2-tools==0.25.17
tinycss2==1.5.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trajectory-msgs==4.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unique-identifier-msgs==2.2.1
uritemplate==3.0.1
uritools==5.0.0
urllib3==1.26.5
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==4.9.0
wadllib==1.3.6
watchdog==6.0.0
webencodings==0.5.1
wrapt==1.13.3
xdg==5
xhtml2pdf==0.2.17
xkit==0.0.0
zipp==1.0.0
